Fix: performace boost
