+++ package.json
@@ -14,6 +14,7 @@
     "@react-navigation/native-stack": "^6.7.0",
     "react": "17.0.2",
     "react-native": "0.68.2",
+    "react-native-gesture-handler": "^2.5.0",
     "react-native-safe-area-context": "^4.3.1",
     "react-native-screens": "^3.15.0",
     "styled-components": "^5.3.5"
