--- src/components/PokemonCard/types.ts
@@ -1,10 +1,0 @@
-import {PokemonDTO} from '../../@types/dtos/pokemonDTO';
-
-export interface Props {
-  item: PokemonDTO;
-}
-
-export interface ColorProps {
-  backgroundColor?: string;
-  isLeft?: boolean;
-}
