+++ src/screens/Home/localComponents/PokemonCard/types.ts
@@ -1,0 +1,10 @@
+import {PokemonDTO} from '../../../../@types/dtos/pokemonDTO';
+
+export type Props = {
+  item: PokemonDTO | undefined;
+};
+
+export interface ColorProps {
+  backgroundColor?: string;
+  isLeft?: boolean;
+}
