+++ src/screens/Home/styles.ts
@@ -1,8 +1,8 @@
 import styled from 'styled-components/native';
 
 export const Container = styled.View`
-  flex: 1;
-  justify-content: center;
-  align-items: center;
+  height: 100%;
+  width: 100%;
   background-color: #f8f8f8;
+  padding: 20px;
 `;
